Main

Description

Descrição : Módulo de Haskell referente à terceira tarefa do projeto de LI. Copyright : Miguel Quaresma miguelquaresma97@gmail.com João Nogueira

Módulo que recebe um ficheiro contendo um mapa, coordenadas e uma linha com um caracter referente à direção pretendida do movimento do boneco, devolvendo, caso este movimento seja válido, as novas coordenadas do boneco ou, caso contrário, as coordenadas iniciais do boneco.

Documentation

main :: IO ()

inStr :: String -> [String]

outStr :: [String] -> String

tarefa3 :: [String] -> [String]

correTestes :: IO ()

Corre todos os testes de um determinado diretório automaticamente, recorrendo à função correTeste

correTeste :: ([String] -> [String]) -> String -> IO ()

Corre um teste e compara o resultado obtido com o esperado

dividemapa :: [String] -> ([String], [String])

Separa o tabuleiro das coordenadas e da letra referente ao movimento do boneco

devolveFinal :: (Int, Int) -> Char -> [String] -> String

Devolve as coordenadas finais do boneco, variando estas consoante o movimento pretendido seja válido ou não

movimentoValido :: [String] -> Char -> (Int, Int) -> Bool

Verifica se o movimento é valido

movimento :: String -> Char

Devolve o caracter que define o movimento

move :: (Int, Int) -> Char -> (Int, Int)

Consoante o caracter presente no fim do ficheiro devolve a coordenada resultante de mover o boneco nessa direção

colocaCaixas :: [String] -> [(Int, Int)] -> [String]

Coloca as caixas no tabuleiro, facilitando assim a validação do movimento do boneco

colocaCaixa :: [String] -> (Int, Int) -> [String]

Função auxiliar da função __colocaCaixas__ que coloca cada caixa individualmente

podeMexerCaixa :: [String] -> Char -> (Int, Int) -> Bool

Verifica se a caixa que o boneco está a empurrar se pode mexer na direção pretendida

processacoordenadas :: [String] -> [(Int, Int)]

Converte a lista de coordenadas, ainda em lista de strings, numa lista de pares de Ints

removeInv :: [String] -> [String]

Remove a coordenadas inválidas (com um so nro; com caracteres que não são nros ou linhas vazias)

remaux :: String -> Bool

removeV :: [String] -> [String]

Remove as linhas vazias da segunda parte do mapa, eliminando possíveis erros de execução