Main

Description

Descrição : Jogo Sokoban escrito em Haskel Copyright : Miguel Quaresma miguelquaresma97@gmail.com João Nogueira joaonogueira097@hotmail.com

Módulo responsável pela implementação do jogo em si, e pelo desenvolvimento da interface do mesmo.

Documentation

type Coordenadas = (Float, Float)

type Mapa = ((Float, Float), Coordenadas, [Coordenadas], [Coordenadas], [Coordenadas], (Int, String), Picture, Picture, Picture, Picture, String, Picture)

Estado do jogo:

Dimensões do mapa * Coordenadas da boneco no mapa * Coordenadas das caixas * Coordenadas das paredes * Coordenadas das posições finais das caixas * Nro de moves * Imagem da boneco * Imagem da caixa * Imagem da parede * Imagem do local de arrumação das caixas * Indicador do estado do mapa * Background do nro de movimentos * Imagem de congratulação

main :: IO ()

gameManager :: Picture -> Picture -> Picture -> Picture -> String -> String -> String -> Picture -> IO ()

Função encarregue de iniciar o jogo

joga :: mundo -> (mundo -> Picture) -> (Event -> mundo -> mundo) -> IO ()

Função que cria um jogo.

desenhaMapa :: Mapa -> Picture

Desenha o jogo dentro da janela

reageManager :: Mapa -> [Mapa] -> Event -> Mapa -> Mapa

Reage a diversos eventos

moveBoneco :: (Float, Float) -> Mapa -> Mapa

Move a boneco uma coordenada para o lado

reageEvento :: Event -> Mapa -> Mapa

Reage ao pressionar das setas do teclado, movendo a bola 5 pixéis numa direção

reageTempo :: Float -> mundo -> mundo

Não reage ao passar do tempo.

move :: (Float, Float) -> Char -> (Float, Float)

Devolve as coords de um determinado um objeto após se movimentar

movimentoValido :: Char -> (Float, Float) -> [Coordenadas] -> [Coordenadas] -> Bool

Verifica se o boneco se pode mover na direção desejada

recolheParedes :: [String] -> Float -> [Coordenadas]

Recolhe as posições dos cardinais/paredes do mapa

recolhePosF :: [String] -> Float -> [Coordenadas]

Recolhe as posições das caixas do mapa

devolveNovo :: Coordenadas -> Char -> [Coordenadas] -> [Coordenadas]

Devolve a lista de coords das caixas com as suas novas posições

listaFinal :: [Coordenadas] -> [Coordenadas] -> Bool

Verifica se todas as caixas estão na posição final

tarefa2 :: [String] -> [String]

Funções da Tarefa 2

simplificaMapa :: [String] -> [(Int, Int)] -> [String] -> [String]

Remove os cardinais redundantes do mapa inteiro

removeDeLinha :: String -> [(Int, Int)] -> [String] -> String

Remove os cardinais redundates de cada linha, verificando, caso o caracter em questão seja um cardinal, todos os caracteres à sua volta e atuando em conformidade

devolveCarater :: (Int, Int) -> [String] -> Char

Devolve o carater correspondente a um par de coordenadas

recolheCords :: Int -> Int -> [String] -> [(Int, Int)]

Recolhe todas as coordenadas referentes a todas posições possíveis no mapa para facilitar a remoção dos cardinais

colocaTudo :: [String] -> [(Int, Int)] -> [String]

Coloca o boneco e as caixas no mapa

colocaBoneco :: [String] -> (Int, Int) -> [String]

Coloca o boneco no mapa

colocaCaixas :: [String] -> [(Int, Int)] -> [String]

Chama a função __colocaCaixa__ para colocar todas as caixas no mapa

colocaCaixa :: [String] -> (Int, Int) -> [String]

Coloca uma caixa no mapa

dividemapa :: [String] -> ([String], [String])

Funções Gerais

Divide o ficheiro em duas partes a primeira contendo o mapa e a segunda as coordenadas

processacoordenadas :: [String] -> [(Int, Int)]

Converte a lista de coordenadas, ainda em lista de strings, numa lista de pares de Ints

removeInv :: [String] -> [String]

Remove a coordenadas inválidas (com um so nro; com caracteres que não são nros ou linhas vazias)

remaux :: String -> Bool

inStr :: String -> [String]