Main

Description

Descrição : Módulo de Haskell referente à quinta tarefa do projeto de LI. Copyright : Miguel Quaresma miguelquaresma97@gmail.com João Nogueira joaonogueira097@hotmail.com

Módulo que recebe uma imagem e calcula a devolve a altura e lagura do menor retângulo onde essa mesma imagem cabe.

Documentation

main :: IO ()

teste :: Picture -> IO ()

tarefa5 :: Picture -> (Float, Float)

medidas :: Picture -> (Float, Float)

Devolve as medidas do retangulo envolvente de uma Picture

pontosFig :: Picture -> Path

Recolhe os pontos principais de uma figura, ou de um conjunto de figuras

rotatePts :: Float -> Path -> Path

Devolve os pontos após um rotação de x graus

pontosCirculo :: Float -> Path

Devolve as coordenadas de pontos do circulo, de 5 em 5 graus